Add support for categorizing switched items
Previously for global switching to work item titles had to match.
This was restrictive if you wanted items to switch together but also
to have different titles. An example of this is when writing
Gradle-related documention where a user may be using the Kotlin or
Groovy DSL and writing settings.gradle.kts and build.gradle.kts or
settings.gradle and build.gradle files respectively.

This commit adds support for categorizing a switched section via its
role. In addition to the role indicating if the section is primary or
secondary, it can now also configure the category by providing
a value, separated from primary or seconday by a space, that is
prefixed with category-.